refactor: updated folder structure
