--- apps/journeys-admin/src/components/PageWrapper/NavigationDrawer/DrawerContent/index.ts
@@ -1,1 +1,0 @@
-export { DrawerContent } from './DrawerContent'
